feat(plugin): adapter plugin -- dremio
