--- plugin/datacap-jdbc-dremio/src/main/resources/META-INF/services/io.edurt.datacap.spi.PluginModule
@@ -1,1 +1,0 @@
-io.edurt.datacap.plugin.jdbc.dremio.DremioPluginModule
