+++ plugin/datacap-plugin-dremio/src/main/java/io/edurt/datacap/plugin/DremioPlugin.java
@@ -1,0 +1,5 @@
+package io.edurt.datacap.plugin;
+
+public class DremioPlugin
+        extends Plugin
+{}
