+++ pom.xml
@@ -37,7 +37,7 @@
         <module>plugin/datacap-plugin-hdfs</module>
         <module>plugin/datacap-plugin-db2</module>
         <module>plugin/datacap-plugin-dm</module>
-        <!--        <module>plugin/datacap-jdbc-dremio</module>-->
+        <module>plugin/datacap-plugin-dremio</module>
         <!--        <module>plugin/datacap-jdbc-druid</module>-->
         <!--        <module>plugin/datacap-jdbc-duckdb</module>-->
         <!--        <module>plugin/datacap-jdbc-elasticsearch</module>-->
